Fix CI/CD
