+++ .vscode/settings.json
@@ -2,9 +2,9 @@
     "python.testing.unittestArgs": [
         "-v",
         "-s",
-        "./tests",
+        ".",
         "-p",
-        "test_*.py"
+        "test*.py"
     ],
     "python.testing.pytestEnabled": false,
     "python.testing.unittestEnabled": true
